+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/epik_code/Blue_Board.java
@@ -137,7 +137,7 @@ public class Blue_Board extends LinearOpMode {
                 })
                 .addTemporalMarker(5, () -> {
                     slide.setPower(-0.5);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(2)
                 .lineTo(new Vector2d(56,58))
@@ -186,7 +186,7 @@ public class Blue_Board extends LinearOpMode {
                 })
                 .addTemporalMarker(5, () -> {
                     slide.setPower(-0.5);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(2)
                 .lineTo(new Vector2d(56,58))
@@ -239,7 +239,7 @@ public class Blue_Board extends LinearOpMode {
                 })
                 .addTemporalMarker(5, () -> {
                     slide.setPower(-0.5);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(2)
                 .lineTo(new Vector2d(56,58))
