+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/epik_code/Red_Audience_Extra_White.java
@@ -174,7 +174,7 @@ public class Red_Audience_Extra_White extends LinearOpMode {
                 })
                 .addTemporalMarker(4, () -> {
                     slide.setPower(-0.7);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(1.5)
                 .lineTo(new Vector2d(56,-6))
@@ -261,7 +261,7 @@ public class Red_Audience_Extra_White extends LinearOpMode {
                 })
                 .addTemporalMarker(4, () -> {
                     slide.setPower(-0.7);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(3)
                 .lineTo(new Vector2d(54,-10))
@@ -351,7 +351,7 @@ public class Red_Audience_Extra_White extends LinearOpMode {
                 })
                 .addTemporalMarker(4, () -> {
                     slide.setPower(-0.7);
-                    slide.setTargetPosition((int) (0));
+                    slide.setTargetPosition(0);
                 })
                 .waitSeconds(1)
                 .lineTo(new Vector2d(56,-5))
